Add stdio support
