+++ test/lib.js
@@ -1,0 +1,6 @@
+const tape = require('tape');
+
+tape.test('todo', test => {
+  test.plan(1);
+  test.equal(true, true, 'todo');
+});
